Unify debug and main commands
