+++ src/bin/main.rs
@@ -1,31 +1,88 @@
 use std::fs;
 
 use fpt::Gameboy;
+use fpt::DebuggerTextInterface;
 
-use clap::Parser;
+use clap::{Parser, Subcommand};
+
+use rustyline::error::ReadlineError;
+use rustyline::{DefaultEditor, Result};
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
 struct Args {
-    rom: String,
     /// Flag to active debug output
     #[arg(short, long)]
     debug: bool,
+
+    #[command(subcommand)]
+    command: Commands,
 }
 
-fn main() {
-    let args = Args::parse();
+#[derive(Subcommand, Debug)]
+enum Commands {
+    /// debugger
+    Debug {},
+    Run {
+        rom: String,
+    }
+}
 
+fn debug() -> Result<()> {
+    let mut debugger_interface = DebuggerTextInterface::new();
+
+    let mut rl = DefaultEditor::new()?;
+    if rl.load_history(".fpt_debug_history").is_err() {
+        println!("No previous history.");
+    }
+    loop {
+        let readline = rl.readline(">> ");
+        match readline {
+            Ok(line) => {
+                let cmd = String::from("return ") + &line;
+                rl.add_history_entry(&line)?;
+                debugger_interface.run(line);
+            }
+            Err(ReadlineError::Interrupted) => {
+                println!("CTRL-C");
+                break;
+            }
+            Err(ReadlineError::Eof) => {
+                println!("CTRL-D");
+                break;
+            }
+            Err(err) => {
+                println!("Error: {:?}", err);
+                break;
+            }
+        }
+    }
+    rl.save_history(".fpt_debug_history")?;
+    Ok(())
+}
+
+
+fn run(rom: String) {
     let mut gameboy = Gameboy::new();
-    gameboy.set_debug(args.debug);
+    //gameboy.set_debug(args.debug);
 
-    let rom = fs::read(args.rom).unwrap();
+    let rom = fs::read(rom).unwrap();
     gameboy.load_rom(&rom);
 
     loop {
-        if args.debug {
-            println!("pc: {:#02X}", gameboy.cpu().pc());
-        }
+        //if args.debug {
+        // println!("pc: {:#02X}", gameboy.cpu().pc());
+        //}
         gameboy.step();
     }
+}
+fn main() {
+    let args = Args::parse();
+
+    match args.command {
+        Commands::Debug{} => { debug();},
+        Commands::Run{rom} => {run(rom);},
+        _ => panic!(),
+    }
+
 }
